osint_app/osint-ingestion/requirements.txt
# Core ML/NLP packages
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Clustering
hdbscan>=0.8.33

# Translation
argostranslate>=1.9.0

# Social media scraping
snscrape>=0.7.0

# Database
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# Image/Media processing
Pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10

# Audio processing (for Whisper)
openai-whisper>=20230918

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional: Performance improvements
faiss-cpu>=1.7.4  # For faster similarity search
numba>=0.57.0     # For JIT compilation